Add battery percent, fix types and printfs

+++ src/main.h
@@ -83,17 +83,17 @@ bool sentDiscovery = false; // Have we successfully sent discovery
 String offline = "offline";
 
 String mqttHost, mqttUser, mqttPass;
-int mqttPort;
+uint16_t mqttPort;
 String room, id, statusTopic, teleTopic, roomsTopic, setTopic;
 
 bool autoUpdate, otaUpdate, prerelease;
 bool discovery, activeScan, publishTele, publishRooms, publishDevices;
-int pirPin, radarPin;
+uint8_t pirPin, radarPin;
 int lastPirValue = -1, lastRadarValue = -1;
 
 #ifdef SENSORS
-int dht11Pin;
-int dht22Pin;
+uint8_t dht11Pin;
+uint8_t dht22Pin;
 
 /** Initialize DHT sensor 1 */
 DHTesp dhtSensor;
